feat(hmac signature verification)
